Extracted common code from ExecutionEnvironmentServices into an abstract class
