+++ packages/controllers/src/services/WebWorkerExecutionEnvironmentService.ts
@@ -1,20 +1,17 @@
-import { Duplex } from 'stream';
+import { ObservableStore } from '@metamask/obs-store';
+import { WorkerParentPostMessageStream } from '@metamask/post-message-stream';
+import { ServiceMessenger } from '@metamask/snap-types';
+import { SNAP_STREAM_NAMES } from '@metamask/snap-workers';
+import { JsonRpcEngine } from 'json-rpc-engine';
+import { createStreamMiddleware } from 'json-rpc-middleware-stream';
 import { nanoid } from 'nanoid';
 import pump from 'pump';
-import { ObservableStore } from '@metamask/obs-store';
-import ObjectMultiplex from '@metamask/object-multiplex';
-import { WorkerParentPostMessageStream } from '@metamask/post-message-stream';
-import { SNAP_STREAM_NAMES } from '@metamask/snap-workers';
-import { createStreamMiddleware } from 'json-rpc-middleware-stream';
-import { SnapExecutionData, ServiceMessenger } from '@metamask/snap-types';
+import { Duplex } from 'stream';
 import {
-  JsonRpcEngine,
-  JsonRpcRequest,
-  PendingJsonRpcResponse,
-} from 'json-rpc-engine';
-import { ExecutionEnvironmentService } from './ExecutionEnvironmentService';
-
-export type SetupSnapProvider = (snapId: string, stream: Duplex) => void;
+  AbstractExecutionService,
+  setupMultiplex,
+  SetupSnapProvider,
+} from './AbstractExecutionService';
 
 interface WorkerControllerArgs {
   setupSnapProvider: SetupSnapProvider;
@@ -30,12 +27,6 @@ interface WorkerStreams {
   _connection: WorkerParentPostMessageStream;
 }
 
-// The snap is the callee
-export type SnapRpcHook = (
-  origin: string,
-  request: Record<string, unknown>,
-) => Promise<unknown>;
-
 interface WorkerWrapper {
   id: string;
   streams: WorkerStreams;
@@ -43,30 +34,10 @@ interface WorkerWrapper {
   worker: Worker;
 }
 
-export class WebWorkerExecutionEnvironmentService
-  implements ExecutionEnvironmentService
-{
+export class WebWorkerExecutionEnvironmentService extends AbstractExecutionService<WorkerWrapper> {
   public store: ObservableStore<{ workers: Record<string, WorkerWrapper> }>;
 
-  private _snapRpcHooks: Map<string, SnapRpcHook>;
-
   private workerUrl: URL;
-
-  private workers: Map<string, WorkerWrapper>;
-
-  private setupSnapProvider: SetupSnapProvider;
-
-  private snapToWorkerMap: Map<string, string>;
-
-  private workerToSnapMap: Map<string, string>;
-
-  private _messenger: ServiceMessenger;
-
-  private _unresponsivePollingInterval: number;
-
-  private _unresponsiveTimeout: number;
-
-  private _timeoutForUnresponsiveMap: Map<string, number>;
 
   constructor({
     setupSnapProvider,
@@ -75,21 +46,18 @@ export class WebWorkerExecutionEnvironmentService
     unresponsivePollingInterval = 5000,
     unresponsiveTimeout = 30000,
   }: WorkerControllerArgs) {
+    super({
+      setupSnapProvider,
+      messenger,
+      unresponsivePollingInterval,
+      unresponsiveTimeout,
+    });
     this.workerUrl = workerUrl;
-    this.setupSnapProvider = setupSnapProvider;
     this.store = new ObservableStore({ workers: {} });
-    this.workers = new Map();
-    this.snapToWorkerMap = new Map();
-    this.workerToSnapMap = new Map();
-    this._snapRpcHooks = new Map();
-    this._messenger = messenger;
-    this._unresponsivePollingInterval = unresponsivePollingInterval;
-    this._unresponsiveTimeout = unresponsiveTimeout;
-    this._timeoutForUnresponsiveMap = new Map();
   }
 
   private _setWorker(workerId: string, workerWrapper: WorkerWrapper): void {
-    this.workers.set(workerId, workerWrapper);
+    this.jobs.set(workerId, workerWrapper);
 
     const newWorkerState = {
       ...(this.store.getState().workers as Record<string, WorkerWrapper>),
@@ -99,7 +67,7 @@ export class WebWorkerExecutionEnvironmentService
   }
 
   private _deleteWorker(workerId: string): void {
-    this.workers.delete(workerId);
+    this.jobs.delete(workerId);
 
     const newWorkerState = {
       ...(this.store.getState().workers as Record<string, WorkerWrapper>),
@@ -108,49 +76,14 @@ export class WebWorkerExecutionEnvironmentService
     this.store.updateState({ workers: newWorkerState });
   }
 
-  private async _command(
-    workerId: string,
-    message: JsonRpcRequest<unknown>,
-  ): Promise<unknown> {
-    if (typeof message !== 'object') {
-      throw new Error('Must send object.');
-    }
-
-    const workerWrapper = this.workers.get(workerId);
-    if (!workerWrapper) {
-      throw new Error(`Worker with id ${workerId} not found.`);
-    }
-
-    console.log('Parent: Sending Command', message);
-    const response: PendingJsonRpcResponse<unknown> =
-      await workerWrapper.rpcEngine.handle(message);
-    if (response.error) {
-      throw new Error(response.error.message);
-    }
-    return response.result;
-  }
-
-  async terminateAllSnaps() {
-    for (const workerId of this.workers.keys()) {
-      this.terminate(workerId);
-    }
-    this._snapRpcHooks.clear();
-  }
-
-  async terminateSnap(snapId: string) {
-    const workerId = this.snapToWorkerMap.get(snapId);
-    workerId && this.terminate(workerId);
-    this._removeSnapHooks(snapId);
-  }
-
   terminate(workerId: string): void {
-    const workerWrapper = this.workers.get(workerId);
+    const workerWrapper = this.jobs.get(workerId);
 
     if (!workerWrapper) {
       throw new Error(`Worker with id "${workerId}" not found.`);
     }
 
-    const snapId = this._getSnapForWorker(workerId);
+    const snapId = this._getSnapForJob(workerId);
 
     if (!snapId) {
       throw new Error(
@@ -167,7 +100,7 @@ export class WebWorkerExecutionEnvironmentService
       }
     });
     workerWrapper.worker.terminate();
-    this._removeSnapAndWorkerMapping(workerId);
+    this._removeSnapAndJobMapping(workerId);
     this._deleteWorker(workerId);
 
     clearTimeout(this._timeoutForUnresponsiveMap.get(workerId));
@@ -176,138 +109,7 @@ export class WebWorkerExecutionEnvironmentService
     console.log(`worker:${workerId} terminated`);
   }
 
-  /**
-   * Gets the RPC message handler for the given snap.
-   *
-   * @param snapId - The id of the Snap whose message handler to get.
-   */
-  async getRpcMessageHandler(snapId: string) {
-    return this._snapRpcHooks.get(snapId);
-  }
-
-  private _removeSnapHooks(snapId: string) {
-    this._snapRpcHooks.delete(snapId);
-  }
-
-  private _createSnapHooks(snapId: string, workerId: string) {
-    const rpcHook = async (
-      origin: string,
-      request: Record<string, unknown>,
-    ) => {
-      return await this._command(workerId, {
-        id: nanoid(),
-        jsonrpc: '2.0',
-        method: 'snapRpc',
-        params: {
-          origin,
-          request,
-          target: snapId,
-        },
-      });
-    };
-
-    this._snapRpcHooks.set(snapId, rpcHook);
-  }
-
-  async executeSnap(snapData: SnapExecutionData): Promise<unknown> {
-    if (this.snapToWorkerMap.has(snapData.snapId)) {
-      throw new Error(`Snap "${snapData.snapId}" is already being executed.`);
-    }
-
-    const worker = await this._initWorker();
-    this._mapSnapAndWorker(snapData.snapId, worker.id);
-    this.setupSnapProvider(
-      snapData.snapId,
-      worker.streams.rpc as unknown as Duplex,
-    );
-
-    const result = await this._command(worker.id, {
-      jsonrpc: '2.0',
-      method: 'executeSnap',
-      params: snapData,
-      id: nanoid(),
-    });
-    // set up poll/ping for status to see if its up, if its not then emit event that it cant be reached
-    this._pollForWorkerStatus(snapData.snapId);
-    this._createSnapHooks(snapData.snapId, worker.id);
-    return result;
-  }
-
-  _pollForWorkerStatus(snapId: string) {
-    const workerId = this._getWorkerForSnap(snapId);
-    if (!workerId) {
-      throw new Error('no worker id found for snap');
-    }
-
-    const timeout = setTimeout(async () => {
-      this._getWorkerStatus(workerId)
-        .then(() => {
-          this._pollForWorkerStatus(snapId);
-        })
-        .catch(() => {
-          this._messenger.publish('ServiceMessenger:unresponsive', snapId);
-        });
-    }, this._unresponsivePollingInterval) as unknown as number;
-    this._timeoutForUnresponsiveMap.set(snapId, timeout);
-  }
-
-  async _getWorkerStatus(workerId: string) {
-    let resolve: any;
-    let reject: any;
-
-    const timeoutPromise = new Promise((res, rej) => {
-      resolve = res;
-      reject = rej;
-    });
-
-    const timeout = setTimeout(() => {
-      reject(new Error('ping request timed out'));
-    }, this._unresponsiveTimeout);
-
-    return Promise.race([
-      this._command(workerId, {
-        jsonrpc: '2.0',
-        method: 'ping',
-        params: [],
-        id: nanoid(),
-      }).then(() => {
-        clearTimeout(timeout);
-        resolve();
-      }),
-      timeoutPromise,
-    ]);
-  }
-
-  _mapSnapAndWorker(snapId: string, workerId: string): void {
-    this.snapToWorkerMap.set(snapId, workerId);
-    this.workerToSnapMap.set(workerId, snapId);
-  }
-
-  /**
-   * @returns The ID of the snap's worker.
-   */
-  _getWorkerForSnap(snapId: string): string | undefined {
-    return this.snapToWorkerMap.get(snapId);
-  }
-
-  /**
-   * @returns The ID worker's snap.
-   */
-  _getSnapForWorker(workerId: string): string | undefined {
-    return this.workerToSnapMap.get(workerId);
-  }
-
-  _removeSnapAndWorkerMapping(workerId: string): void {
-    const snapId = this.workerToSnapMap.get(workerId);
-    if (!snapId) {
-      throw new Error(`worker:${workerId} has no mapped snap.`);
-    }
-
-    this.workerToSnapMap.delete(workerId);
-    this.snapToWorkerMap.delete(snapId);
-  }
-
-  private async _initWorker(): Promise<WorkerWrapper> {
+  protected async _initJob(): Promise<WorkerWrapper> {
     const workerId = nanoid();
     const worker = new Worker(this.workerUrl, {
       name: workerId,
@@ -315,7 +117,7 @@ export class WebWorkerExecutionEnvironmentService
     // Handle out-of-band errors, i.e. errors thrown from the snap outside of the req/res cycle.
     const errorHandler = (ev: ErrorEvent) => {
       if (this._messenger) {
-        const snapId = this.workerToSnapMap.get(workerId);
+        const snapId = this.jobToSnapMap.get(workerId);
         if (snapId) {
           this._messenger.publish('ServiceMessenger:unhandledError', snapId, {
             code: ev.error.code,
@@ -371,32 +173,4 @@ export class WebWorkerExecutionEnvironmentService
       _connection: workerStream,
     };
   }
-}
-
-/**
- * Sets up stream multiplexing for the given stream.
- *
- * @param connectionStream - the stream to mux
- * @param streamName - the name of the stream, for identification in errors
- * @return {stream.Stream} the multiplexed stream
- */
-function setupMultiplex(
-  connectionStream: Duplex,
-  streamName: string,
-): ObjectMultiplex {
-  const mux = new ObjectMultiplex();
-  pump(
-    connectionStream,
-    // Typecast: stream type mismatch
-    mux as unknown as Duplex,
-    connectionStream,
-    (err) => {
-      if (err) {
-        streamName
-          ? console.error(`${streamName} stream failure.`, err)
-          : console.error(err);
-      }
-    },
-  );
-  return mux;
 }
