+++ packages/controllers/src/services/index.ts
@@ -1,2 +1,3 @@
-export * from './ExecutionEnvironmentService';
+export * from './AbstractExecutionService';
+export * from './ExecutionService';
 export * from './WebWorkerExecutionEnvironmentService';
